Factor out `<MainSplitter>`
